requirements.txt
django
django-dashing
djangorestframework